Refactor(performance): Split the profiler module
profiler.py held 657 lines, past the 400-line reviewability limit the
aislop code-quality gate applies.

Split it into a package at the same dotted path: models (the category
enum and the metric records), tracking (the psutil-backed
OperationTimer and MemoryTracker plus the profile_operation decorator)
and profiler (PerformanceProfiler with ProfileReport). The last two
stay together because ProfileReport annotates PerformanceProfiler
unquoted, so separating them would mean editing that annotation rather
than moving code. The package __init__ re-exports the previous names.
